refactor(board-health): split the dispatch-gate tests into their own module
This PR grew board_health_dispatch_gate.rs from 771 to 1617 lines
(71325 bytes), over both Server Guards ceilings (1500 lines / 51200
bytes). Move the test module into board_health_dispatch_gate_tests.rs
behind #[path], the convention repo_graph_generation.rs and run_dirs.rs
already use in this crate.

Source drops to 973 lines / 44741 bytes and tests to 644 / 24209, both
within the limits. No test or production logic changed; 25 tests pass.
